+++ lib/widgets/hymn/hymn_verses_section.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/material.dart';
+import '../../models/hymn.dart';
+import 'hymn_detail_widgets.dart';
+
+class HymnVersesSection extends StatelessWidget {
+  final Hymn hymn;
+  final double fontSize;
+  final double countFontSize;
+
+  const HymnVersesSection({
+    super.key,
+    required this.hymn,
+    required this.fontSize,
+    required this.countFontSize,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      children: [
+        for (int i = 0; i < hymn.verses.length; i++)
+          HymnVerseDisplayWidget(
+            verseNumber: i + 1,
+            verseText: hymn.verses[i],
+            fontSize: fontSize,
+            countFontSize: countFontSize,
+          ),
+      ],
+    );
+  }
+}
