+++ lib/widgets/player/modern_audio_player_widget.dart
@@ -5,6 +5,12 @@ import 'package:just_audio/just_audio.dart';
 import '../../models/hymn.dart';
 import 'package:fihirana/services/audio/audio_service.dart';
 import '../../l10n/app_localizations.dart';
+import 'player_top_bar.dart';
+import 'album_art_card.dart';
+import 'song_info_section.dart';
+import 'progress_slider.dart';
+import 'player_controls.dart';
+import 'playlist_bottom_sheet.dart';
 
 class ModernAudioPlayerWidget extends StatefulWidget {
   final Hymn hymn;
@@ -262,20 +268,12 @@ class _ModernAudioPlayerWidgetState extends State<ModernAudioPlayerWidget> {
     super.dispose();
   }
 
-  String _formatDuration(Duration duration) {
-    String twoDigits(int n) => n.toString().padLeft(2, '0');
-    final minutes = twoDigits(duration.inMinutes.remainder(60));
-    final seconds = twoDigits(duration.inSeconds.remainder(60));
-    return '$minutes:$seconds';
-  }
+
 
   @override
   Widget build(BuildContext context) {
     // Dark theme colors based on the image
     const backgroundColor = Color(0xFF1C1B1F); // Dark background
-    const cardColor = Color(0xFFE0C09C); // Beige/Gold card color from image
-    const primaryTextColor = Colors.white;
-    const secondaryTextColor = Colors.white70;
 
     return Container(
       color: backgroundColor,
@@ -300,83 +298,15 @@ class _ModernAudioPlayerWidgetState extends State<ModernAudioPlayerWidget> {
             child: Column(
               children: [
                 // Top Bar
-                Padding(
-                  padding:
-                      const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
-                  child: Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                    children: [
-                      IconButton(
-                        icon: const Icon(Icons.keyboard_arrow_down,
-                            color: Colors.white, size: 30),
-                        onPressed:
-                            widget.onClose ?? () => Navigator.of(context).pop(),
-                      ),
-                      Text(
-                        'Now Playing',
-                        style: TextStyle(
-                          color: primaryTextColor.withValues(alpha: 0.9),
-                          fontSize: 14,
-                          fontWeight: FontWeight.w600,
-                          letterSpacing: 1,
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
+                PlayerTopBar(onClose: widget.onClose),
 
                 const SizedBox(height: 20),
 
                 // Album Art Card
                 Expanded(
                   flex: 5,
-                  child: Padding(
-                    padding: const EdgeInsets.symmetric(horizontal: 32),
-                    child: Container(
-                      decoration: BoxDecoration(
-                        color: cardColor,
-                        borderRadius: BorderRadius.circular(30),
-                        boxShadow: [
-                          BoxShadow(
-                            color: Colors.black.withValues(alpha: 0.3),
-                            blurRadius: 20,
-                            offset: const Offset(0, 10),
-                          ),
-                        ],
-                      ),
-                      child: Stack(
-                        children: [
-                          // Placeholder for Hymn Image (or just the color/gradient)
-                          Center(
-                            child: Icon(
-                              Icons.music_note,
-                              size: 120,
-                              color: Colors.brown.withValues(alpha: 0.3),
-                            ),
-                          ),
-                          // Hymn Number Overlay
-                          Positioned(
-                            top: 20,
-                            right: 20,
-                            child: Container(
-                              padding: const EdgeInsets.symmetric(
-                                  horizontal: 12, vertical: 6),
-                              decoration: BoxDecoration(
-                                color: Colors.black.withValues(alpha: 0.1),
-                                borderRadius: BorderRadius.circular(20),
-                              ),
-                              child: Text(
-                                '#${_currentDisplayedHymn?.hymnNumber ?? widget.hymn.hymnNumber}',
-                                style: TextStyle(
-                                  color: Colors.brown.shade900,
-                                  fontWeight: FontWeight.bold,
-                                ),
-                              ),
-                            ),
-                          ),
-                        ],
-                      ),
-                    ),
+                  child: AlbumArtCard(
+                    hymnNumber: _currentDisplayedHymn?.hymnNumber ?? widget.hymn.hymnNumber,
                   ),
                 ),
 
@@ -391,179 +321,42 @@ class _ModernAudioPlayerWidgetState extends State<ModernAudioPlayerWidget> {
                       crossAxisAlignment: CrossAxisAlignment.start,
                       children: [
                         // Title and Heart
-                        Row(
-                          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                          crossAxisAlignment: CrossAxisAlignment.center,
-                          children: [
-                            Expanded(
-                              child: Column(
-                                crossAxisAlignment: CrossAxisAlignment.start,
-                                children: [
-                                   Text(
-                                     _currentDisplayedHymn?.title ?? widget.hymn.title,
-                                     style: const TextStyle(
-                                       color: primaryTextColor,
-                                       fontSize: 24,
-                                       fontWeight: FontWeight.bold,
-                                     ),
-                                     maxLines: 1,
-                                     overflow: TextOverflow.ellipsis,
-                                   ),
-                                   const SizedBox(height: 4),
-                                   Text(
-                                     'Hymn ${_currentDisplayedHymn?.hymnNumber ?? widget.hymn.hymnNumber}',
-                                     style: const TextStyle(
-                                       color: secondaryTextColor,
-                                       fontSize: 16,
-                                     ),
-                                     maxLines: 1,
-                                     overflow: TextOverflow.ellipsis,
-                                   ),
-                                ],
-                              ),
-                            ),
-                            Container(
-                              decoration: BoxDecoration(
-                                shape: BoxShape.circle,
-                                border:
-                                    Border.all(color: Colors.white24, width: 1),
-                              ),
-                              child: IconButton(
-                                icon: Icon(
-                                  widget.isFavorite
-                                      ? Icons.favorite
-                                      : Icons.favorite_border,
-                                  color: widget.isFavorite
-                                      ? Colors.red
-                                      : Colors.white,
-                                ),
-                                onPressed: widget.onToggleFavorite,
-                              ),
-                            ),
-                          ],
+                        SongInfoSection(
+                          title: _currentDisplayedHymn?.title ?? widget.hymn.title,
+                          hymnNumber: _currentDisplayedHymn?.hymnNumber ?? widget.hymn.hymnNumber,
+                          isFavorite: widget.isFavorite,
+                          onToggleFavorite: widget.onToggleFavorite,
                         ),
 
                         const SizedBox(height: 30),
 
                         // Progress Slider
-                        SliderTheme(
-                          data: SliderTheme.of(context).copyWith(
-                            activeTrackColor: Colors.white,
-                            inactiveTrackColor: Colors.white24,
-                            thumbColor: Colors.white,
-                            trackHeight: 2,
-                            thumbShape: const RoundSliderThumbShape(
-                                enabledThumbRadius: 6),
-                            overlayShape: const RoundSliderOverlayShape(
-                                overlayRadius: 14),
-                          ),
-                          child: Slider(
-                            value: _isDraggingSlider
-                                ? _dragValue
-                                : (_duration != null &&
-                                        _position != null &&
-                                        _duration!.inMilliseconds > 0)
-                                    ? (_position!.inMilliseconds.toDouble() /
-                                            _duration!.inMilliseconds
-                                                .toDouble())
-                                        .clamp(0.0, 1.0)
-                                    : 0.0,
-                            onChanged: (value) {
-                              setState(() {
-                                _isDraggingSlider = true;
-                                _dragValue = value;
-                              });
-                            },
-                            onChangeEnd: _seekTo,
-                          ),
-                        ),
-
-                        // Time Indicators
-                        Padding(
-                          padding: const EdgeInsets.symmetric(horizontal: 6),
-                          child: Row(
-                            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                            children: [
-                              Text(
-                                _formatDuration(_isDraggingSlider
-                                    ? Duration(
-                                        milliseconds: (_dragValue *
-                                                (_duration?.inMilliseconds ??
-                                                    0))
-                                            .round())
-                                    : _position ?? Duration.zero),
-                                style: const TextStyle(
-                                    color: secondaryTextColor, fontSize: 12),
-                              ),
-                              Text(
-                                _formatDuration(_duration ?? Duration.zero),
-                                style: const TextStyle(
-                                    color: secondaryTextColor, fontSize: 12),
-                              ),
-                            ],
-                          ),
+                        ProgressSlider(
+                          duration: _duration,
+                          position: _position,
+                          isDragging: _isDraggingSlider,
+                          dragValue: _dragValue,
+                          onChanged: (value) {
+                            setState(() {
+                              _isDraggingSlider = true;
+                              _dragValue = value;
+                            });
+                          },
+                          onChangeEnd: _seekTo,
                         ),
 
                         const SizedBox(height: 20),
 
                         // Main Controls
-                        Row(
-                          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                          children: [
-                            Switch(
-                              value: widget.autoPlayNext,
-                              onChanged: widget.onAutoPlayNextChange,
-                              activeThumbColor: Colors.white,
-                              activeTrackColor: Colors.lightBlue,
-                              inactiveThumbColor: Colors.white54,
-                              inactiveTrackColor: Colors.white10,
-
-                            ),
-                            IconButton(
-                              icon: const Icon(Icons.skip_previous_rounded,
-                                  color: Colors.white, size: 32),
-                              onPressed: _playPrevious,
-                            ),
-
-                            // Play/Pause (Large White Circle)
-                            GestureDetector(
-                              onTap: _togglePlayPause,
-                              child: Container(
-                                width: 70,
-                                height: 70,
-                                decoration: const BoxDecoration(
-                                  color: Colors.white,
-                                  shape: BoxShape.circle,
-                                ),
-                                child: _isLoading
-                                    ? const Padding(
-                                        padding: EdgeInsets.all(20),
-                                        child: CircularProgressIndicator(
-                                          color: Colors.black,
-                                          strokeWidth: 3,
-                                        ),
-                                      )
-                                    : Icon(
-                                        _isPlaying
-                                            ? Icons.pause
-                                            : Icons.play_arrow,
-                                        color: Colors.black,
-                                        size: 36,
-                                      ),
-                              ),
-                            ),
-
-                            // Next (30 Sec Forward style)
-                            IconButton(
-                              icon: const Icon(Icons.skip_next_rounded,
-                                  color: Colors.white, size: 32),
-                              onPressed: _playNext,
-                            ),
-                            IconButton(
-                              icon: const Icon(Icons.playlist_add_check_circle, color: Colors.white),
-                              onPressed: () => _showPlaylist(context),
-                            ),
-                          ],
+                        PlayerControls(
+                          isPlaying: _isPlaying,
+                          isLoading: _isLoading,
+                          autoPlayNext: widget.autoPlayNext,
+                          onAutoPlayNextChange: widget.onAutoPlayNextChange,
+                          onPlayPrevious: _playPrevious,
+                          onTogglePlayPause: _togglePlayPause,
+                          onPlayNext: _playNext,
+                          onShowPlaylist: () => _showPlaylist(context),
                         ),
                       ],
                     ),
@@ -585,139 +378,13 @@ class _ModernAudioPlayerWidgetState extends State<ModernAudioPlayerWidget> {
         borderRadius: BorderRadius.vertical(top: Radius.circular(20)),
       ),
       builder: (context) {
-        return StatefulBuilder(
-          builder: (context, setState) {
-            return Container(
-              padding: const EdgeInsets.symmetric(vertical: 16),
-              child: Column(
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  Padding(
-                    padding: const EdgeInsets.symmetric(horizontal: 16),
-                    child: Row(
-                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                      children: [
-                        const Text(
-                          'Playlist',
-                          style: TextStyle(
-                            color: Colors.white,
-                            fontSize: 18,
-                            fontWeight: FontWeight.bold,
-                          ),
-                        ),
-                        IconButton(
-                          icon: const Icon(Icons.close, color: Colors.white),
-                          onPressed: () => Navigator.pop(context),
-                        ),
-                      ],
-                    ),
-                  ),
-                  const Divider(color: Colors.white24),
-                  Expanded(
-                    child: widget.playlist == null || widget.playlist!.isEmpty
-                        ? const Center(
-                            child: Text(
-                              'No hymns in playlist',
-                              style: TextStyle(color: Colors.white54),
-                            ),
-                          )
-                        : ListView.builder(
-                            itemCount: widget.playlist!.length,
-                            itemBuilder: (context, index) {
-                              final hymn = widget.playlist![index];
-                               final isCurrent =
-                                   hymn.id == _currentDisplayedHymn?.id &&
-                                   !hymn.id.startsWith('recording_');
-                              final isDownloaded =
-                                  widget.isDownloaded?.call(hymn) ?? false;
-                              final downloadProgress =
-                                  widget.getDownloadProgress?.call(hymn);
-
-                              return ListTile(
-                                selected: isCurrent,
-                                selectedTileColor: Colors.white10,
-                                leading: Container(
-                                  width: 40,
-                                  height: 40,
-                                  decoration: BoxDecoration(
-                                    color: isCurrent
-                                        ? Colors.white
-                                        : Colors.white10,
-                                    shape: BoxShape.circle,
-                                  ),
-                                  child: Center(
-                                    child: isCurrent
-                                        ? const Icon(Icons.bar_chart,
-                                            color: Colors.black, size: 20)
-                                        : Text(
-                                            hymn.hymnNumber,
-                                            style: const TextStyle(
-                                              color: Colors.white,
-                                              fontWeight: FontWeight.bold,
-                                            ),
-                                          ),
-                                  ),
-                                ),
-                                title: Text(
-                                  hymn.title,
-                                  style: TextStyle(
-                                    color: isCurrent
-                                        ? Colors.white
-                                        : Colors.white70,
-                                    fontWeight: isCurrent
-                                        ? FontWeight.bold
-                                        : FontWeight.normal,
-                                  ),
-                                  maxLines: 1,
-                                  overflow: TextOverflow.ellipsis,
-                                ),
-                                subtitle: Text(
-                                  'Hymn ${hymn.hymnNumber}',
-                                  style: const TextStyle(
-                                    color: Colors.white54,
-                                    fontSize: 12,
-                                  ),
-                                ),
-                                trailing: Row(
-                                  mainAxisSize: MainAxisSize.min,
-                                  children: [
-                                    if (downloadProgress != null)
-                                      SizedBox(
-                                        width: 24,
-                                        height: 24,
-                                        child: CircularProgressIndicator(
-                                          value: downloadProgress,
-                                          strokeWidth: 2,
-                                          color: Colors.white,
-                                        ),
-                                      )
-                                    else if (isDownloaded)
-                                      const Icon(Icons.check_circle,
-                                          color: Colors.green, size: 20)
-                                    else
-                                      IconButton(
-                                        icon: const Icon(Icons.download_rounded,
-                                            color: Colors.white54),
-                                        onPressed: () {
-                                          widget.onDownload?.call(hymn);
-                                          setState(
-                                              () {}); // Refresh to show progress if immediate
-                                        },
-                                      ),
-                                  ],
-                                ),
-                                onTap: () {
-                                  widget.onHymnChange?.call(hymn);
-                                  Navigator.pop(context);
-                                },
-                              );
-                            },
-                          ),
-                  ),
-                ],
-              ),
-            );
-          },
+        return PlaylistBottomSheet(
+          playlist: widget.playlist,
+          currentHymn: _currentDisplayedHymn,
+          onHymnChange: widget.onHymnChange,
+          onDownload: widget.onDownload,
+          isDownloaded: widget.isDownloaded,
+          getDownloadProgress: widget.getDownloadProgress,
         );
       },
     );
