+++ lib/widgets/recording/recording_upload_dialog.dart
@@ -1,0 +1,193 @@
+import 'package:flutter/material.dart';
+import 'package:get/get.dart';
+import '../../controller/recording_controller.dart';
+import '../../controller/color_controller.dart';
+import '../../l10n/app_localizations.dart';
+import '../../models/user_recording.dart';
+
+class RecordingUploadDialog extends StatelessWidget {
+  final RecordingController controller;
+  final ColorController colorController;
+  final UserRecording recording;
+  final VoidCallback onDone;
+
+  const RecordingUploadDialog({
+    super.key,
+    required this.controller,
+    required this.colorController,
+    required this.recording,
+    required this.onDone,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    final l10n = AppLocalizations.of(Get.context!)!;
+
+    return Dialog(
+      backgroundColor: Colors.transparent,
+      child: Container(
+        margin: EdgeInsets.symmetric(
+          horizontal: MediaQuery.of(Get.context!).size.width * 0.15,
+          vertical: MediaQuery.of(Get.context!).size.height * 0.2,
+        ),
+        decoration: BoxDecoration(
+          gradient: LinearGradient(
+            begin: Alignment.topCenter,
+            end: Alignment.bottomCenter,
+            colors: [
+              colorController.primaryColor.value,
+              colorController.primaryColor.value.withValues(alpha: 0.8),
+            ],
+          ),
+          borderRadius: BorderRadius.circular(20),
+          boxShadow: [
+            BoxShadow(
+              color: Colors.black.withValues(alpha: 0.3),
+              blurRadius: 20,
+              offset: const Offset(0, 10),
+            ),
+          ],
+        ),
+        child: ConstrainedBox(
+          constraints: BoxConstraints(
+            maxHeight: MediaQuery.of(Get.context!).size.height * 0.5,
+            maxWidth: MediaQuery.of(Get.context!).size.width * 0.7,
+          ),
+          child: SingleChildScrollView(
+            padding: const EdgeInsets.all(24),
+            child: Column(
+              mainAxisSize: MainAxisSize.min,
+              children: [
+                const Row(
+                  children: [
+                    Icon(
+                      Icons.cloud_upload,
+                      color: Colors.white,
+                      size: 24,
+                    ),
+                    SizedBox(width: 12),
+                    Expanded(
+                      child: Text(
+                        'Uploading to Drive',
+                        style: TextStyle(
+                          color: Colors.white,
+                          fontSize: 18,
+                          fontWeight: FontWeight.bold,
+                        ),
+                      ),
+                    ),
+                  ],
+                ),
+                const SizedBox(height: 20),
+                Obx(() {
+                  final isUploading =
+                      controller.isUploadingRecording(recording.id);
+                  final uploadError =
+                      controller.getUploadError(recording.id);
+
+                  if (isUploading) {
+                    return Column(
+                      children: [
+                        const CircularProgressIndicator(
+                          valueColor:
+                              AlwaysStoppedAnimation<Color>(Colors.white),
+                        ),
+                        const SizedBox(height: 16),
+                        Text(
+                          'Uploading...',
+                          style: TextStyle(
+                            color: Colors.white.withValues(alpha: 0.8),
+                          ),
+                        ),
+                      ],
+                    );
+                  } else if (uploadError != null) {
+                    return Column(
+                      children: [
+                        const Icon(Icons.error_outline,
+                            color: Colors.redAccent, size: 48),
+                        const SizedBox(height: 16),
+                        const Text(
+                          'Upload Failed',
+                          style: TextStyle(
+                            color: Colors.white,
+                            fontWeight: FontWeight.bold,
+                          ),
+                        ),
+                        const SizedBox(height: 8),
+                        Text(
+                          uploadError,
+                          style: TextStyle(
+                            color: Colors.white.withValues(alpha: 0.8),
+                            fontSize: 12,
+                          ),
+                          textAlign: TextAlign.center,
+                        ),
+                        const SizedBox(height: 20),
+                        Row(
+                          children: [
+                            Expanded(
+                              child: ElevatedButton(
+                                onPressed: () {
+                                  controller.retryUpload(recording);
+                                },
+                                style: ElevatedButton.styleFrom(
+                                  backgroundColor: Colors.white,
+                                  foregroundColor:
+                                      colorController.primaryColor.value,
+                                ),
+                                child: Text(l10n.retry),
+                              ),
+                            ),
+                            const SizedBox(width: 12),
+                            Expanded(
+                              child: ElevatedButton(
+                                onPressed: onDone,
+                                style: ElevatedButton.styleFrom(
+                                  backgroundColor:
+                                      Colors.white.withValues(alpha: 0.2),
+                                  foregroundColor: Colors.white,
+                                ),
+                                child: Text(l10n.skip),
+                              ),
+                            ),
+                          ],
+                        ),
+                      ],
+                    );
+                  } else {
+                    // Success
+                    return Column(
+                      children: [
+                        const Icon(Icons.check_circle_outline,
+                            color: Colors.white, size: 48),
+                        const SizedBox(height: 16),
+                        const Text(
+                          'Upload Complete!',
+                          style: TextStyle(
+                            color: Colors.white,
+                            fontWeight: FontWeight.bold,
+                          ),
+                        ),
+                        const SizedBox(height: 20),
+                        ElevatedButton(
+                          onPressed: onDone,
+                          style: ElevatedButton.styleFrom(
+                            backgroundColor: Colors.white,
+                            foregroundColor:
+                                colorController.primaryColor.value,
+                          ),
+                          child: Text(l10n.done),
+                        ),
+                      ],
+                    );
+                  }
+                }),
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
